have goal list show summary of first activity if the goal description is blank
Before when an activity was added to a non-goal, the goal description would be
set to 'See Action Plan'.  Now the description will be left blank, and the
goal_list page will pull the description of the first activity instead.  If
there are no activities yet then it will show 'See Action Plan'