change in question parsing of multi checkbox questions
Multiple checkboxes only pass a value back to the server when they are checked,
so have to do some tweaking to get this right.  First thing is the checkboxes
need a value.  Then that value has to be coded into the question, and the
question parser.  Then the questions parser can try to match the answer to the
value of the checkbox to know if it should be checked.