delay empty goal creation until subgoal is saved
When adding a subgoal before creating it's parent goal, delay actual creation
of the parent goal until the subgoal is being saved.

This way a goal isn't created until the subgoal is created.  Otherwise a person
may click to add a subgoal, and change their mind, but still end up with an
empty goal.