dont forward goal description now that we aren't using it
The goal list uses either the goal description or the first sub-goal description
Now that the goal description isn't being used, we shouldn't forward the goal
description from previous years.  If we do there is no way to change what is
on the goal list because the goal description is shown or editable.