go ahead and destroy the session if the token doesn't match
Mostly this is because someone's ip address changed, but they couldn't log back
in.  With the session cleaned they should at least be able to get back in.

There is the potential for a DoS by someone connecting with various session ids,
but they would pretty much have to be sniffing the network to do that
effectively.  So I'm not worried about it.