Updates to the framework and silence some syslog noise
Proper guessing of the base_url, which uses a Longest Common String algorithm
ported from my perl framework.
Guessing of the Content-Type if the page pushes all the page in the first
argument to output() without setting the third argument.  This is for AJAX
stuff. (Not that csip does AJAX.)
Bunch of php warnings about Undefined variables and indexes silenced.