Adding more security - user courses
If the user has an allowed course list, use it to limit their access.

Remove the report permissions fro plc members, so they only see their course.
remove the save button unless this is a plc member in one of their courses.
save_answer and delete_cfa check that the course is in the user's allowed
courses list.
updated cfa_new.php to allow it to hide the save button also.