Add handling of multiple answers to a question for results tab
Add answerid to all questions, and all questions account for the answer being
an array instead of a scalar.  All parts but the Results tab just grab the
first element of the answers array though.  Modified course_save_answer() to
need the answerid, with there being more than one answer to a question I
can't guess the answerid anymore.
The Results tab always had a blank set of questions at the bottom for more
answers.