Add modules with read/write settings sets to database (and ldap) config
Also a few other tweaks to bring the framework up-to-date with another
project.  Most notably the addition of the ldap handle wrapper class.

After this I need to add support for getting user info from an external
source (PowerSchool).  And after that build in an automatic info sync
for user,location, and course info.  Lots to do still.