Add v8 Form - #1 Version the Form and add questions to the database
Also fixed a little data problem with extra parts and answers by adding a
unique key to the database.
Force the questionids instead of letting AUTO_INCREMENT handle it.  This
makes sense where the questionid is hard coded in some parts of the form.
Add the v8 questions to the database with an update script and in the install
file.
Still need to test versioning of the cfa_new script.