Fix new cfa - part number collision because of earlier deleted part
The new cfa was saying part already existed, because it did.  The count was
wrong because of a missing number in the part numbers (and they were out of
order).  So the number for the next part is based on part numbers now.  And
I switched the html element id's from count to part number too.