Add a whole bunch of database error logging
Still having reports of trouble in the results tab.  This adds a whole bunch
of error log statements for database errors.  Also changes the method for
getting a new sequence value when the sequence isn't in the sequence table,
uses max() instead of count(), and polls all questions in the group with a
group by.