v8 form: principal_comment, dynamic version in new_year, recycle question groups
Add principal_comment to the course_approval table, to load_csip, and to
approve.php, with a function in inc/courses.phpm for setting it.
In manage/new_year.php make the version dynamic based on what's available in
the database.  So when I add the v8 questions version 8 will be available
automatically.
recycle the question_group numbers in v8 questions, so that I don't have to
add more rows to the course_question_links table.