Adding external sync for users, locations, and courses.
Add external id's to those three tables.  And a hardlink map table in the
config file.  Also add a database update script.
Add external id display to the edit pages for those three tables.
Add tie-in for OAuth and modulare authentication to lib/security.phpm
Add PS.phpm to handle the PowerSchool API side of things.  This is the primary
source for external info.
Add to the install.php the ability to pre-sync align those three tables (set
the external ids).