Switch lib/PS (auth->external module for PowerSchool) to OpenID
Switch to the JanRain OpenID module, which is packaged in Ubuntu 14.04.
Fix some syntax and reference errors.
Add the option of a login through external button.  This must have the
'external_auth_button' id attribute on the link/button/submit tag.  Which will
cause it to be styled similar to other buttons in CSIP.
With this auth->external and external user sync through PowerSchool API is
almost done.  I've added an option to separate user/location sync from
course sync, but haven't implemented that in the PS.phpm module.