Pull get_token() out of the lib/PS constructor so login page is faster
The login page was firing two http requests to PowerSchool because of this,
so the login page could take a while to appear.  This moves the get_token()
call to later in the process, actually to right before the token is needed,
so that the login page can show up quickly.