PS Integration Pre-Sync - Add an explicit label to the CSIP side for courses
I've added the option of an explicit label on the CSIP side of the matching
per item.  This is so that I can have a complex value for the course label,
instead of just the name.  We decided it would help to have the course
category name and grade levels too.