With a new sequence make sure all questions on the group get updates
The auto-save needs to make sure all questions in a repeatable group get the
new sequence value when the first question in a new group gets saved.
And the new sequence value should never be '0'.
And make sure that an empty string in answers_original can be found and
removed.  So that the javascript knows the answer hasn't actually changed
when it hasn't.