Add a field to the School Leadership course questions for Tech Plan
This means adding a question group, and setting part 1 of that course to use
it, with the added text and question input.
I had to extend the clean_question() / mangle_question() functions to handle
the utot_url and utot_label config values, since they are part of the
question that's usually on the Team PGP Tab (part 1).

I'll probably have to change the text and input placeholder later.